refactor: extract AgentHistoryContentValue to a standalone webapps-schema type
Pulls the content-block record out of AgentHistoryEntity so it can also be
reused by AgentInstanceEntity's systemPrompt field.
